% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hic200.R
\name{bins_for_region}
\alias{bins_for_region}
\title{Convert a genomic region to genome-wide bin indices}
\usage{
bins_for_region(index, chr, start, end, res = 200L)
}
\arguments{
\item{index}{Output of \code{\link[=hic200_index]{hic200_index()}}.}

\item{chr}{Chromosome name.}

\item{start, end}{Region start/end in bp.}

\item{res}{Bin size in bp.}
}
\value{
Integer vector \code{c(bin0, bin1)} (inclusive).
}
\description{
Convert a genomic region to genome-wide bin indices
}
