% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{drawBW}
\alias{drawBW}
\title{Draw a signal track}
\usage{
drawBW(
  hic,
  dataName = "data1",
  type = "area",
  fill = "#354863",
  fill2 = NULL,
  col = "#354863",
  xaxis = TRUE,
  xtitle = TRUE,
  ymin = NULL,
  ymax = NULL,
  ylabel = NULL,
  unit = 1e+06,
  ntick = 3
)
}
\arguments{
\item{hic}{A Hi-C object.}

\item{dataName}{Name of an attached track.}

\item{type}{One of \code{"area"}, \code{"BorderStrength"}, \code{"tad"}, \code{"line"}.}

\item{fill, fill2}{Fill colours for positive/negative values.}

\item{col}{Line colour (for \code{type = "line"}).}

\item{xaxis, xtitle}{Show x axis text / title.}

\item{ymin, ymax}{Y-axis limits.}

\item{ylabel}{Strip label (defaults to \code{dataName}).}

\item{unit}{x-axis unit.}

\item{ntick}{Approximate number of y ticks.}
}
\value{
A ggplot object.
}
\description{
Draw a signal track
}
