% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{drawGene}
\alias{drawGene}
\title{Draw the gene model track}
\usage{
drawGene(hic, label = FALSE, unit = 1e+06, xaxis = FALSE)
}
\arguments{
\item{hic}{A Hi-C object.}

\item{label}{Draw gene symbols?}

\item{unit}{x-axis unit.}

\item{xaxis}{Show x axis text?}
}
\value{
A ggplot object.
}
\description{
Draw the gene model track
}
