% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicdraw-package.R
\docType{package}
\name{hicdraw-package}
\alias{hicdraw}
\alias{hicdraw-package}
\title{hicdraw: Drawing and Handling of Hi-C Contact Maps and Domain Boundaries}
\description{
Tools for reading, slicing and plotting Hi-C contact maps together with boundary-strength and domain annotation tracks. Supports the sparse genome-wide bin matrix produced by hic200-cpp as well as the .hic format read through 'strawr'. A configuration-file driven interface lets the user generate many figures in a single call.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rafysta/hicdraw}
  \item Report bugs at \url{https://github.com/rafysta/hicdraw/issues}
}

}
\author{
\strong{Maintainer}: Hideki Noma \email{rafysta@gmail.com}

Authors:
\itemize{
  \item Hideki Noma \email{rafysta@gmail.com}
}

}
\keyword{internal}
