% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_object.R
\name{setRegion}
\alias{setRegion}
\title{Set the genomic region to draw}
\usage{
setRegion(hic, chr, start, end)
}
\arguments{
\item{hic}{A Hi-C object from \code{\link[=createHiCmap]{createHiCmap()}}.}

\item{chr}{Chromosome name.}

\item{start, end}{Region in bp.}
}
\value{
The updated Hi-C object.
}
\description{
Set the genomic region to draw
}
